VB-4785 - Optimise notification service
- Move methods into their own services
- Remove unused classes / code
- Split tests into their own files (1 file per event type)
- Add handling of errors at client level to remove use of try catch at service level
