+++ src/main/kotlin/uk/gov/justice/digital/hmpps/notificationsalertsvsip/service/NotificationService.kt
@@ -2,17 +2,13 @@ package uk.gov.justice.digital.hmpps.notificationsalertsvsip.service
 
 import org.slf4j.Logger
 import org.slf4j.LoggerFactory
-import org.springframework.http.HttpStatus
 import org.springframework.stereotype.Service
-import org.springframework.web.reactive.function.client.WebClientResponseException
-import uk.gov.justice.digital.hmpps.notificationsalertsvsip.client.VisitSchedulerClient
-import uk.gov.justice.digital.hmpps.notificationsalertsvsip.dto.visit.scheduler.VisitDto
 import uk.gov.justice.digital.hmpps.notificationsalertsvsip.enums.VisitEventType
 import java.time.LocalDateTime
 
 @Service
 class NotificationService(
-  val visitSchedulerClient: VisitSchedulerClient,
+  val visitSchedulerService: VisitSchedulerService,
   val smsSenderService: SmsSenderService,
 ) {
   private companion object {
@@ -20,29 +16,15 @@ class NotificationService(
   }
 
   fun sendMessage(visitEventType: VisitEventType, bookingReference: String) {
-    LOG.info("Visit booked event with reference - $bookingReference")
+    LOG.info("Received call to send notification for event type $visitEventType, visit - $bookingReference")
 
-    val visit = getVisit(bookingReference)
-
+    val visit = visitSchedulerService.getVisit(bookingReference)
     visit?.let {
       if (visit.startTimestamp > LocalDateTime.now() && !visit.visitContact.telephone.isNullOrEmpty()) {
         smsSenderService.sendSms(visit, visitEventType)
       } else {
         LOG.info("Visit in past or no telephone number exists for contact on visit reference - ${visit.reference}")
       }
-    }
-  }
-
-  private fun getVisit(bookingReference: String): VisitDto? {
-    try {
-      return visitSchedulerClient.getVisitByReference(bookingReference)
-    } catch (e: Exception) {
-      // if there was an error getting the visit return null
-      if (e is WebClientResponseException && e.statusCode == HttpStatus.NOT_FOUND) {
-        LOG.error("no visit found with booking reference - $bookingReference")
-        return null
-      }
-      throw e
     }
   }
 }
